+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Changelog for NeoFS Node
 - IR's estimations timer management (#3539)
 
 ### Updated
+- `github.com/nspcc-dev/neofs-sdk-go` dependency to `v1.0.0-rc.14.0.20250909131532-07fa82695f26` (#3569)
 
 ### Updating from v0.48.3
 Drop `--main-only` flag from all `neofs-cli object head` commands.